requirements.txt
opencv-python
Pillow
numpy
pandas
easyocr
pyzbar
ultralytics